Refactor the structure of NameData related store and manager
