+++ src/components/ndn-lib/src/lib.rs
@@ -3,8 +3,7 @@
 mod chunk;
 mod object;
 mod link_obj;
-mod local_store;
-mod named_data_mgr;
+mod named_data;
 mod cyfs_http;
 mod ndn_client;
 mod fileobj;
@@ -16,9 +15,8 @@ mod object_array;
 
 pub use object::*;
 pub use chunk::*;
-pub use local_store::*;
 pub use link_obj::*;
-pub use named_data_mgr::*;
+pub use named_data::*;
 pub use cyfs_http::*;
 pub use ndn_client::*;
 pub use fileobj::*;
