+++ src/components/ndn-lib/src/named_data/mod.rs
@@ -1,0 +1,19 @@
+mod def;
+mod named_data_db;
+mod named_data_mgr;
+mod named_data_mgr_db;
+mod named_data_store;
+
+pub use def::*;
+pub use named_data_db::*;
+pub use named_data_mgr::*;
+pub use named_data_mgr_db::*;
+pub use named_data_store::*;
+
+#[cfg(test)]
+mod test_store;
+
+#[cfg(test)]
+mod test_mgr;
+
+
